LPS-202905 Generate the link ASAP to avoid performance issues
